Trial for improvement on new logger : move code related to specific platform into the right header
